+++ internal/daemon/db/entity/notifications_queue.go
@@ -1,0 +1,6 @@
+package entity
+
+type NotificationsQueue struct {
+	Subject string `json:"Subject"`
+	Body    string `json:"Body"`
+}