+++ internal/setting/analyzer/login.go
@@ -1,5 +1,9 @@
 package analyzer
 
+import (
+	"git.kor-elf.net/kor-elf-shield/kor-elf-shield/internal/daemon/analyzer/config"
+)
+
 type Login struct {
 	Enabled bool `mapstructure:"enabled"`
 	Notify  bool `mapstructure:"notify"`
@@ -38,4 +42,30 @@ func defaultLogin() Login {
 
 func (l Login) Validate() error {
 	return nil
+}
+
+func (l Login) ToSources() ([]*config.Source, error) {
+	var sources []*config.Source
+
+	if !l.Enabled {
+		return sources, nil
+	}
+
+	if l.SSHEnable {
+		sources = append(sources, config.NewLoginSSH(l.Notify && l.SSHNotify)...)
+	}
+
+	if l.LocalEnable {
+		sources = append(sources, config.NewLoginLocal(l.Notify && l.LocalNotify)...)
+	}
+
+	if l.SuEnable {
+		sources = append(sources, config.NewLoginSu(l.Notify && l.SuNotify)...)
+	}
+
+	if l.SudoEnable {
+		sources = append(sources, config.NewLoginSudo(l.Notify && l.SudoNotify)...)
+	}
+
+	return sources, nil
 }